fix autodownload for oxparis and minidepth. Remove angle-field decoder from inner timings if not activated in joint_point_line_extractor.py

+++ gluefactory/datasets/oxford_paris_mini_1view_jpldd.py
@@ -1,14 +1,12 @@
 import logging
-import os
 import pickle
 import random
 import shutil
-import tarfile
+import zipfile
 from pathlib import Path
 
 import numpy as np
 import torch
-from tqdm import tqdm
 
 from gluefactory.datasets import BaseDataset, augmentations
 from gluefactory.settings import DATA_PATH, root
@@ -103,7 +101,7 @@ class OxfordParisMiniOneViewJPLDD(BaseDataset):
             "test": images,
             "all": images,
         }
-        print(f"DATASET OVERALL(NO-SPLIT) IMAGES: {len(images)}")
+        logger.info(f"DATASET OVERALL(NO-SPLIT) IMAGES: {len(images)}")
 
         augmentation_map = {
             "dark": augmentations.DarkAugmentation,
@@ -114,40 +112,26 @@ class OxfordParisMiniOneViewJPLDD(BaseDataset):
         self.augmentation = augmentation_map[self.conf.load_features.augment.type]()
 
     def download_oxford_paris_mini(self):
+        """
+        The downloaded dataset already contains ground-truth keypoints, line-df and line-af and lines
+        for dataset original resolution.
+        """
         logger.info("Downloading the OxfordParis Mini dataset...")
-        data_dir = DATA_PATH / self.conf.data_dir
-        tmp_dir = data_dir.parent / "oxpa_tmp"
+        oxparis_root_directory = (DATA_PATH / self.conf.data_dir).parent
+        tmp_dir = oxparis_root_directory.parent / "oxpa_tmp"
         if tmp_dir.exists():
             shutil.rmtree(tmp_dir)
         tmp_dir.mkdir(exist_ok=True, parents=True)
-        url_base = "http://ptak.felk.cvut.cz/revisitop/revisitop1m/"
-        num_parts = 100
-        # go through dataset parts, one by one and only keep wanted images in img_list
-        for i in tqdm(range(num_parts), position=1):
-            tar_name = f"revisitop1m.{i + 1}.tar.gz"
-            tar_url = url_base + "jpg/" + tar_name
-            tmp_tar_path = tmp_dir / tar_name
-            torch.hub.download_url_to_file(tar_url, tmp_tar_path)
-            with tarfile.open(tmp_tar_path) as tar:
-                tar.extractall(path=data_dir)
-            tmp_tar_path.unlink()
-            # Delete unwanted files
-            existing_files = set(
-                [str(i.relative_to(data_dir)) for i in data_dir.glob("**/*.jpg")]
-            )
-            to_del = existing_files - set(self.img_list)
-            for d in to_del:
-                Path(data_dir / d).unlink()
-
-        shutil.rmtree(tmp_dir)
-
-        # remove empty directories
-        for file in os.listdir(data_dir):
-            cur_file: Path = data_dir / file
-            if cur_file.is_file():
-                continue
-            if len(os.listdir(cur_file)) == 0:
-                shutil.rmtree(cur_file)
+        url = "https://filedn.com/lt6zb4ORSwapNyVniJf1Pqh/JPL/oxparis_gt_complete.zip"
+        zip_name = "oxparis.zip"
+        zip_path = tmp_dir / zip_name
+        torch.hub.download_url_to_file(url, zip_path)
+        with zipfile.ZipFile(zip_path, "r") as zip_ref:
+            zip_ref.extractall(tmp_dir)
+        shutil.move(tmp_dir / "revisitop1m_POLD2", str(oxparis_root_directory))
+        logger.info("Delete temporary files...")
+        if tmp_dir.exists():
+            shutil.rmtree(tmp_dir)
 
     def get_dataset(self, split):
         assert split in ["train", "val", "test", "all"]
